Fixed problem displaying mathjax from obsidian
Mathjax on GitHub behaves differently from Obsidian. It's more strict on $ being on a new line to start an expression. Additionally using $ to start an expression seems to cause problems in making spaces inside an expression. This can be solved by using ```math instead of $ but this isn't supported on obsidian. This is why I'm making an exception and committing directly from GitHub.